Create RentalHandle Module
